requirements.txt
# pip install -r requirements.txt

# pytorch ----------------------------------------
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.2
Pillow
PyYAML>=5.3.1
scipy>=1.4.1
torch>=1.7.0
torchvision>=0.8.1
tqdm>=4.41.0
pandas>=1.3.1

# flask ------------------------------------------
flask>=1.1.2
flask-cors>=3.0.10
python-dotenv
gevent>=21.1.2
gunicorn>=20.0.4
requests>=2.26.0
seaborn>=0.11.1
